requirements.txt
requests
websocket_client
./dist/NorenRestApiPy-0.0.16-py2.py3-none-any.whl
pandas
pyyaml
